requirements.txt
# Web
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Config & validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Embeddings
sentence-transformers>=3.0.0

# Vector store
faiss-cpu>=1.7.4
chromadb>=0.4.0

# PDF
pypdf>=4.0.0

# LLM
openai>=1.0.0
httpx>=0.27.0

# Dev / test
pytest>=8.0.0
pytest-asyncio>=0.24.0
